feat: refactor entity classes to implement SoftDeletableEntity and streamline properties
