+++ backend/Domain/Common/BaseEntities/AuditableEntity.cs
@@ -1,10 +1,9 @@
-namespace Domain.Common.Base;
+namespace Domain.Common.BaseEntities;
 
-public class AuditableEntity : CreatedOnlyEntity
+public abstract class AuditableEntity
 {
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+    public Guid? CreatedBy { get; set; }
     public DateTime? UpdatedAt { get; set; }
-    public int? UpdatedBy { get; set; }
-    public bool IsDeleted { get; set; } = false;
-    public DateTime? DeletedAt { get; set; }
-    public int? DeletedBy { get; set; }
+    public Guid? UpdatedBy { get; set; }
 }
