+++ backend/Domain/Common/BaseEntities/SoftDeletableEntity.cs
@@ -1,6 +1,8 @@
 ﻿namespace Domain.Common.BaseEntities;
 
-public class SoftDeletableEntity
+public abstract class SoftDeletableEntity : AuditableEntity
 {
-    
+    public DateTime? DeletedAt { get; set; }
+    public Guid? DeletedBy { get; set; }
+    public bool IsDeleted { get; set; } = false;
 }
