+++ backend/Domain/Entities/Blog.cs
@@ -1,9 +1,11 @@
-﻿namespace Domain.Entities;
+﻿using Domain.Common.BaseEntities;
+
+namespace Domain.Entities;
 
 /// <summary>
-/// Table for storing blog posts
+///     Table for storing blog posts
 /// </summary>
-public class Blog
+public class Blog : SoftDeletableEntity
 {
     public Guid Id { get; set; }
 
@@ -14,18 +16,6 @@ public class Blog
     public string Author { get; set; } = null!;
 
     public DateTime? PublishedAt { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public Guid? CreatedBy { get; set; }
-
-    public DateTime? UpdatedAt { get; set; }
-
-    public Guid? UpdatedBy { get; set; }
-
-    public DateTime? DeletedAt { get; set; }
-
-    public Guid? DeletedBy { get; set; }
 
     public ICollection<BlogTag> BlogTag { get; set; } = [];
