+++ backend/Domain/Entities/BlogTag.cs
@@ -1,25 +1,15 @@
-﻿namespace Domain.Entities;
+﻿using Domain.Common.BaseEntities;
+
+namespace Domain.Entities;
 
 /// <summary>
-/// Join table for many-to-many relationship between blogs and tags
+///     Join table for many-to-many relationship between blogs and tags
 /// </summary>
-public class BlogTag
+public class BlogTag : SoftDeletableEntity
 {
     public Guid BlogId { get; set; }
 
     public Guid TagId { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public Guid? CreatedBy { get; set; }
-
-    public DateTime? UpdatedAt { get; set; }
-
-    public Guid? UpdatedBy { get; set; }
-
-    public DateTime? DeletedAt { get; set; }
-
-    public Guid? DeletedBy { get; set; }
 
     public Blog Blog { get; set; } = null!;
