+++ backend/Domain/Entities/Message.cs
@@ -1,18 +1,12 @@
-﻿namespace Domain.Entities;
+﻿using Domain.Common.BaseEntities;
 
-public class Message
+namespace Domain.Entities;
+
+public class Message : SoftDeletableEntity
 {
     public Guid Id { get; set; }
 
     public Guid ConversationId { get; set; }
-
-    public Guid CreatedBy { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public Guid? UpdateBy { get; set; }
-
-    public DateTime? UpdatedAt { get; set; }
 
     public string Content { get; set; } = null!;
