+++ backend/Domain/Entities/Purchase.cs
@@ -1,18 +1,12 @@
-﻿namespace Domain.Entities;
+﻿using Domain.Common.BaseEntities;
 
-public class Purchase
+namespace Domain.Entities;
+
+public class Purchase : SoftDeletableEntity
 {
     public Guid Id { get; set; }
 
     public Guid AccountId { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public Guid? CreatedBy { get; set; }
-
-    public DateTime? DeletedAt { get; set; }
-
-    public Guid? DeletedBy { get; set; }
 
     public Account Account { get; set; } = null!;
